requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# LLM APIs
anthropic>=0.40.0
voyageai>=0.2.0

# LangChain ecosystem
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-text-splitters>=0.3.0

# Vector database
qdrant-client>=1.7.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
greenlet>=3.0.0

# Document processing
pymupdf>=1.23.0
python-docx>=1.1.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
aiofiles>=23.2.1
